The Outliers/api/requirements.txt
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.1
Flask==2.3.3
Flask-JWT-Extended==4.5.3
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lightgbm==4.6.0
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.23
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.0.6
